Renamed entry generator
- Entry generator renamed to templates
- readme.md updated with new program flow
